% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutor.R
\name{berakna_mittpunkter}
\alias{berakna_mittpunkter}
\title{Beräkna mittpunktskoordinater för rutor}
\usage{
berakna_mittpunkter(
  df,
  xruta,
  yruta,
  rutstorlek,
  xkolnamn = "mitt_x",
  ykolnamn = "mitt_y"
)
}
\arguments{
\item{df}{En \code{data.frame} med x- och y-kolumner (nedre vänstra hörnet).}

\item{xruta, yruta}{Namn på koordinatkolumnerna.}

\item{rutstorlek}{Rutstorlek i meter.}

\item{xkolnamn, ykolnamn}{Namn på de nya mittpunktskolumnerna.}
}
\value{
\code{df} med två nya kolumner, placerade efter \code{xruta}/\code{yruta}.
}
\description{
Beräkna mittpunktskoordinater för rutor
}
