% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartdata.R
\name{hamta_karttabell}
\alias{hamta_karttabell}
\title{Tabell över kartlager som går att hämta med \code{hamta_karta()}}
\usage{
hamta_karttabell()
}
\value{
En \code{data.frame} med \code{namn}, \code{id_kol}, \code{lankol}, \code{kommunkol} och
\code{sokord} (listkolumn).
}
\description{
Tabell över kartlager som går att hämta med \code{hamta_karta()}
}
