% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer.R
\name{las_gisfil_fran_zipfil_via_sokvag}
\alias{las_gisfil_fran_zipfil_via_sokvag}
\title{Läs en GIS-fil ur en lokal zip-fil}
\usage{
las_gisfil_fran_zipfil_via_sokvag(skickad_sokvag)
}
\arguments{
\item{skickad_sokvag}{Sökväg till zip-filen.}
}
\value{
Ett \code{sf}-objekt.
}
\description{
Läs en GIS-fil ur en lokal zip-fil
}
