% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer.R
\name{las_gisfil_fran_zipfil_via_url}
\alias{las_gisfil_fran_zipfil_via_url}
\title{Läs en GIS-fil ur en zip-fil på en URL}
\usage{
las_gisfil_fran_zipfil_via_url(skickad_url)
}
\arguments{
\item{skickad_url}{URL till zip-filen.}
}
\value{
Ett \code{sf}-objekt.
}
\description{
Laddar ner zip-filen, packar upp den och läser den med \code{sf::st_read()}.
}
