% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartdata.R
\name{postgis_postgistabell_till_sf}
\alias{postgis_postgistabell_till_sf}
\title{Läs en PostGIS-tabell direkt till ett sf-objekt}
\usage{
postgis_postgistabell_till_sf(con = "default", schema, tabell, query = NA)
}
\arguments{
\item{con}{En \code{DBIConnection} eller \code{"default"} (använder
\code{rdpostgres::uppkoppling_db()}).}

\item{schema, tabell}{Schema och tabell.}

\item{query}{Valfri SQL-svans efter \verb{SELECT * FROM schema.tabell}.}
}
\value{
Ett \code{sf}-objekt.
}
\description{
Läs en PostGIS-tabell direkt till ett sf-objekt
}
