% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovrigt.R
\name{raster_till_vektor}
\alias{raster_till_vektor}
\title{Vektorisera ett rasterlager till polygoner}
\usage{
raster_till_vektor(
  rasterlager,
  filtrera_bort_na = TRUE,
  behall_varden_over = 0
)
}
\arguments{
\item{rasterlager}{Ett \code{terra::SpatRaster}.}

\item{filtrera_bort_na}{Ta bort celler med \code{NA}.}

\item{behall_varden_over}{Behåll bara polygoner med värde större än detta
(\code{NA} = ingen filtrering).}
}
\value{
Ett \code{sf}-objekt med en kolumn \code{varde}.
}
\description{
Vektorisera ett rasterlager till polygoner
}
