% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdgis-package.R
\docType{package}
\name{rdgis-package}
\alias{rdgis}
\alias{rdgis-package}
\title{rdgis: GIS-verktyg för Region Dalarnas Analysarbete}
\description{
Hämta kartdata ur geodatabasen, skapa sf-objekt av ruttabeller och koordinatpar, geometriska hjälpfunktioner (centroider inuti polygoner, linjer längs punkter), läsa GIS-filer ur zip-arkiv och extrahera kolumnnamn ur ESRI-geodatabaser. Utbrutet ur geometri-avsnitten i func_GIS.R i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
