% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutor.R
\name{rutstorlek_estimera}
\alias{rutstorlek_estimera}
\title{Gissa rutstorlek från koordinatvärden}
\usage{
rutstorlek_estimera(x, y)
}
\arguments{
\item{x, y}{Numeriska vektorer med koordinater.}
}
\value{
\code{100}, \code{500}, \code{1000} eller \code{NA}.
}
\description{
SCB:s rutkoordinater ligger i nedre vänstra hörnet. Storleken framgår av hur
värdena slutar.
}
