% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutor.R
\name{sf_fran_df_med_x_y_kol}
\alias{sf_fran_df_med_x_y_kol}
\title{Skapa ett sf-objekt från en data.frame med x- och y-kolumner}
\usage{
sf_fran_df_med_x_y_kol(
  skickad_df,
  x_kol,
  y_kol,
  rutstorlek = NA,
  polygonlager = TRUE,
  vald_crs = 3006
)
}
\arguments{
\item{skickad_df}{En \code{data.frame}.}

\item{x_kol, y_kol}{Namn på koordinatkolumnerna (nedre vänstra hörnet).}

\item{rutstorlek}{Rutstorlek i meter (\code{NA} = gissa med
\code{\link[=rutstorlek_estimera]{rutstorlek_estimera()}}).}

\item{polygonlager}{\code{TRUE} = kvadratiska cellpolygoner, \code{FALSE} = punkter.}

\item{vald_crs}{Koordinatsystem (EPSG). Standard 3006 (SWEREF99 TM).}
}
\value{
Ett \code{sf}-objekt.
}
\description{
Skapa ett sf-objekt från en data.frame med x- och y-kolumner
}
