% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometri.R
\name{skapa_linje_langs_med_punkter}
\alias{skapa_linje_langs_med_punkter}
\title{Dra en linje längs en serie punkter i angiven ordning}
\usage{
skapa_linje_langs_med_punkter(
  skickad_sf,
  kol_ord,
  namn_kol,
  namn_bara_startpunkt = TRUE
)
}
\arguments{
\item{skickad_sf}{Ett \code{sf}-objekt med punkter.}

\item{kol_ord}{Namn på en numerisk kolumn som anger ordningen.}

\item{namn_kol}{Namn på kolumn med punktnamn (för etiketter).}

\item{namn_bara_startpunkt}{\code{TRUE} = etikett från startpunkten;
\code{FALSE} = \code{"start - slut"}.}
}
\value{
Ett \code{sf}-objekt med linjesegment (\code{start}, \code{end}, \code{label}).
}
\description{
Dra en linje längs en serie punkter i angiven ordning
}
