% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovrigt.R
\name{skapa_punkt_sf_av_koordinatpar}
\alias{skapa_punkt_sf_av_koordinatpar}
\title{Skapa (eller fyll på) ett sf-objekt av koordinatpar}
\usage{
skapa_punkt_sf_av_koordinatpar(
  koordinater,
  malpunktsnamn,
  sf_obj = NULL,
  vald_crs = 4326
)
}
\arguments{
\item{koordinater}{Teckenvektor med \code{"lat, lon"}-strängar.}

\item{malpunktsnamn}{Namn för varje punkt (samma längd som \code{koordinater}).}

\item{sf_obj}{Befintligt \code{sf}-objekt att lägga till punkterna i (\code{NULL} =
skapa nytt).}

\item{vald_crs}{Koordinatsystem (EPSG). Standard 4326 (WGS84).}
}
\value{
Ett \code{sf}-objekt med punkter (\code{id}, \code{malpunkt_namn}, \code{lat}, \code{lon}).
}
\description{
Skapa (eller fyll på) ett sf-objekt av koordinatpar
}
