% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometri.R
\name{spatial_join_med_ovrkat}
\alias{spatial_join_med_ovrkat}
\title{Spatial join där punkter utanför områdeslagret får ett "övrigt"-värde}
\usage{
spatial_join_med_ovrkat(
  gislager_grunddata,
  gislager_omr,
  omrade_kol = "omrade",
  ovrig_varde = "Övriga områden"
)
}
\arguments{
\item{gislager_grunddata}{\code{sf}-objekt att koda.}

\item{gislager_omr}{\code{sf}-objekt med områdespolygoner.}

\item{omrade_kol}{Namn på områdeskolumnen i \code{gislager_omr}.}

\item{ovrig_varde}{Värde för objekt som hamnar utanför alla områden.}
}
\value{
Ett \code{sf}-objekt.
}
\description{
Spatial join där punkter utanför områdeslagret får ett "övrigt"-värde
}
