% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometri.R
\name{st_centroid_within_geo}
\alias{st_centroid_within_geo}
\title{Centroider som garanterat ligger inuti sin polygon}
\usage{
st_centroid_within_geo(x, ensure_within = TRUE, of_largest_polygon = TRUE)
}
\arguments{
\item{x}{Ett \code{sf}-objekt med polygoner.}

\item{ensure_within}{Tvinga punkten att ligga innanför polygonen
(\code{st_point_on_surface} när den geometriska centroiden hamnar utanför).}

\item{of_largest_polygon}{Använd den största delpolygonen för multipolygoner.}
}
\value{
Ett \code{sf}-objekt med punkter.
}
\description{
Centroider som garanterat ligger inuti sin polygon
}
\seealso{
\url{https://github.com/r-spatial/sf/issues/1302}
}
