% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometri.R
\name{st_largest_ring}
\alias{st_largest_ring}
\title{Största ringen i varje (multi)polygon}
\usage{
st_largest_ring(x)
}
\arguments{
\item{x}{Ett \code{sf}-objekt med polygoner.}
}
\value{
Ett \code{sf}-objekt där varje rad ersatts med sin största delpolygon.
}
\description{
Största ringen i varje (multi)polygon
}
\seealso{
\url{https://github.com/r-spatial/sf/issues/1302}
}
