% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer.R
\name{unzip_zipfil_med_zipfiler}
\alias{unzip_zipfil_med_zipfiler}
\title{Packa upp en zip-fil som själv innehåller zip-filer}
\usage{
unzip_zipfil_med_zipfiler(skickad_url)
}
\arguments{
\item{skickad_url}{URL till den yttre zip-filen.}
}
\value{
En teckenvektor med sökvägar till de uppackade (inre) zip-filerna.
}
\description{
Används t.ex. för FA-/LA-regioner och läns-/kommungränser med kustlinje, som
distribueras som zip-filer i en zip-fil.
}
